Tweaked all the examples. Tried to rename some stuff. Got the henneberg solution working better.
