+++ examples/3_Advanced/planar_coil_optimization/QH_minimal.py
@@ -9,8 +9,6 @@ import time
 import numpy as np
 from scipy.optimize import minimize
 from simsopt.field import BiotSavart, Current, coils_via_symmetries
-# from simsopt.field import CoilCoilNetForces, CoilCoilNetTorques, \
-#     TotalVacuumEnergy, CoilSelfNetForces, CoilCoilNetForces12, CoilCoilNetTorques12
 from simsopt.field import regularization_rect
 from simsopt.field.force import MeanSquaredForce, coil_force, coil_torque, coil_net_torques, coil_net_forces, LpCurveForce, \
     SquaredMeanForce, \
@@ -29,19 +27,16 @@ import re
 
 t1 = time.time()
 
-# Number of Fourier modes describing each Cartesian component of each coil:
-order = 0
-
 # File for the desired boundary magnetic surface:
 TEST_DIR = (Path(__file__).parent / ".." / ".." / "tests" / "test_files").resolve()
-input_name = 'input.LandremanPaul2021_QA_reactorScale_lowres'
+input_name = 'input.LandremanPaul2021_QH_reactorScale_lowres'
 filename = TEST_DIR / input_name
 
 # Initialize the boundary magnetic surface:
 range_param = "half period"
 nphi = 32
 ntheta = 32
-poff = 1.5
+poff = 2.0
 coff = 3.0
 s = SurfaceRZFourier.from_vmec_input(filename, range=range_param, nphi=nphi, ntheta=ntheta)
 s_inner = SurfaceRZFourier.from_vmec_input(filename, range=range_param, nphi=nphi * 4, ntheta=ntheta * 4)
@@ -64,7 +59,7 @@ s_plot = SurfaceRZFourier.from_vmec_input(
 )
 
 ### Initialize some TF coils
-def initialize_coils_QA(TEST_DIR, s):
+def initialize_coils_QH(TEST_DIR, s):
     """
     Initializes coils for each of the target configurations that are
     used for permanent magnet optimization.
@@ -85,28 +80,34 @@ def initialize_coils_QA(TEST_DIR, s):
     from simsopt.geo import curves_to_vtk
 
     # generate planar TF coils
-    ncoils = 3
+    ncoils = 2
     R0 = s.get_rc(0, 0) * 1
-    R1 = s.get_rc(1, 0) * 3
+    R1 = s.get_rc(1, 0) * 4
     order = 8
 
     from simsopt.mhd.vmec import Vmec
-    vmec_file = 'wout_LandremanPaul2021_QA_reactorScale_lowres_reference.nc'
-    total_current = Vmec(TEST_DIR / vmec_file).external_current() / (2 * s.nfp) / 1.2
+    vmec_file = 'wout_LandremanPaul2021_QH_reactorScale_lowres_reference.nc'
+    total_current = Vmec(TEST_DIR / vmec_file).external_current() / (2 * s.nfp) / 1.4
     print('Total current = ', total_current)
 
     # Only need Jax flag for CurvePlanarFourier class
     base_curves = create_equally_spaced_curves(
         ncoils, s.nfp, stellsym=True, 
         R0=R0, R1=R1, order=order, numquadpoints=256,
-        jax_flag=False,
+        jax_flag=True,
     )
 
     base_currents = [(Current(total_current / ncoils * 1e-7) * 1e7) for _ in range(ncoils - 1)]
+    # base_currents = [(Current(total_current / ncoils * 1e-7) * 1e7) for _ in range(ncoils)]
+    # base_currents[0].fix_all()
+
     total_current = Current(total_current)
     total_current.fix_all()
     base_currents += [total_current - sum(base_currents)]
     coils = coils_via_symmetries(base_curves, base_currents, s.nfp, True)
+    # for c in coils:
+    #     c.current.fix_all()
+    #     c.curve.fix_all()
 
     # Initialize the coil curves and save the data to vtk
     curves = [c.curve for c in coils]
@@ -114,8 +115,8 @@ def initialize_coils_QA(TEST_DIR, s):
     return base_curves, curves, coils, base_currents
 
 # initialize the coils
-base_curves_TF, curves_TF, coils_TF, currents_TF = initialize_coils_QA(TEST_DIR, s)
-num_TF_unique_coils = len(coils_TF) // 4
+base_curves_TF, curves_TF, coils_TF, currents_TF = initialize_coils_QH(TEST_DIR, s)
+num_TF_unique_coils = len(base_curves_TF) 
 base_coils_TF = coils_TF[:num_TF_unique_coils]
 currents_TF = np.array([coil.current.get_value() for coil in coils_TF])
 
@@ -129,6 +130,7 @@ calculate_on_axis_B(bs_TF, s)
 # Only need this if make self forces and TVE nonzero in the objective! 
 a = 0.2
 b = 0.2
+nturns = 100
 nturns_TF = 200
 
 def pointData_forces_torques(coils, allcoils, aprimes, bprimes, nturns_list):
@@ -152,23 +154,29 @@ def pointData_forces_torques(coils, allcoils, aprimes, bprimes, nturns_list):
 btot = bs_TF
 calculate_on_axis_B(btot, s)
 btot.set_points(s.gamma().reshape((-1, 3)))
-a_list = np.ones(len(coils_TF)) * a
-b_list = np.ones(len(coils_TF)) * b
+# a_list = np.hstack((np.ones(len(coils)) * aa, np.ones(len(coils_TF)) * a))
+# b_list = np.hstack((np.ones(len(coils)) * bb, np.ones(len(coils_TF)) * b))
+# base_a_list = np.hstack((np.ones(len(base_coils)) * aa, np.ones(len(base_coils_TF)) * a))
+# base_b_list = np.hstack((np.ones(len(base_coils)) * bb, np.ones(len(base_coils_TF)) * b))
 
 LENGTH_WEIGHT = Weight(0.001)
-LENGTH_TARGET = 180
-LINK_WEIGHT = 1e4
+LENGTH_TARGET = 80
+LINK_WEIGHT = 1e3
 CC_THRESHOLD = 0.8
 CC_WEIGHT = 1e1
 CS_THRESHOLD = 1.5
 CS_WEIGHT = 1e2
 # Weight for the Coil Coil forces term
 FORCE_WEIGHT = Weight(0.0) # Forces are in Newtons, and typical values are ~10^5, 10^6 Newtons
-FORCE_WEIGHT2 = Weight(0.0) # Forces are in Newtons, and typical values are ~10^5, 10^6 Newtons
-TORQUE_WEIGHT = Weight(0.0) # Forces are in Newtons, and typical values are ~10^5, 10^6 Newtons
-TORQUE_WEIGHT2 = Weight(0.0) # Forces are in Newtons, and typical values are ~10^5, 10^6 Newtons
+# FORCE_WEIGHT2 = Weight(0.0) # Forces are in Newtons, and typical values are ~10^5, 10^6 Newtons
+# TORQUE_WEIGHT = Weight(1e-24) # Forces are in Newtons, and typical values are ~10^5, 10^6 Newtons
+# TORQUE_WEIGHT2 = Weight(1e-24) # Forces are in Newtons, and typical values are ~10^5, 10^6 Newtons
+# FORCE_WEIGHT = Weight(1e-22) # Forces are in Newtons, and typical values are ~10^5, 10^6 Newtons
+# FORCE_WEIGHT2 = Weight(0.0) # Forces are in Newtons, and typical values are ~10^5, 10^6 Newtons
+# TORQUE_WEIGHT = Weight(1e-24) # Forces are in Newtons, and typical values are ~10^5, 10^6 Newtons
+# TORQUE_WEIGHT2 = Weight(1e-24) # Forces are in Newtons, and typical values are ~10^5, 10^6 Newtons
 # Directory for output
-OUT_DIR = ("./QA_nodipoles_lowerror/")
+OUT_DIR = ("./QH_nodipoles/")
 if os.path.exists(OUT_DIR):
     shutil.rmtree(OUT_DIR)
 os.makedirs(OUT_DIR, exist_ok=True)
@@ -188,6 +196,10 @@ for c in (coils_TF):
 
 pointData = {"B_N": np.sum(btot.B().reshape((nphi, ntheta, 3)) * s.unitnormal(), axis=2)[:, :, None]}
 s.to_vtk(OUT_DIR + "surf_init_DA", extra_data=pointData)
+
+btot.set_points(s_plot.gamma().reshape((-1, 3)))
+pointData = {"B_N": np.sum(btot.B().reshape((qphi, qtheta, 3)) * s_plot.unitnormal(), axis=2)[:, :, None]}
+s_plot.to_vtk(OUT_DIR + "surf_full_init_DA", extra_data=pointData)
 btot.set_points(s.gamma().reshape((-1, 3)))
 
 # Repeat for whole B field
@@ -208,78 +220,124 @@ Jls_TF = [CurveLength(c) for c in base_curves_TF]
 Jlength = QuadraticPenalty(sum(Jls_TF), LENGTH_TARGET, "max")
 
 # coil-coil and coil-plasma distances should be between all coils
+
+### Jcc below removed the dipoles!
 Jccdist = CurveCurveDistance(curves_TF, CC_THRESHOLD, num_basecurves=len(coils_TF))
 Jcsdist = CurveSurfaceDistance(curves_TF, s, CS_THRESHOLD)
+Jcsdist2 = CurveSurfaceDistance(curves_TF, s, CS_THRESHOLD)
 
 # While the coil array is not moving around, they cannot
 # interlink. 
-# While the coil array is not moving around, they cannot
-# interlink. 
-linkNum = LinkingNumber(curves_TF, downsample=2)
+linkNum = LinkingNumber(curves_TF)
 
-# Currently, all force terms involve all the coils
-all_coils = coils_TF
-all_base_coils = base_coils_TF
-Jforce = sum([LpCurveForce(c, all_coils, regularization_rect(a_list[i], b_list[i]), p=4, threshold=4e5 * 100, downsample=1
-    ) for i, c in enumerate(all_base_coils)])
-Jforce2 = sum([SquaredMeanForce(c, all_coils, downsample=1) for c in all_base_coils])
+##### Note need coils_TF + coils below!!!!!!!
+# Jforce2 = sum([LpCurveForce(c, coils_TF, 
+#     regularization=regularization_rect(base_a_list[i], base_b_list[i]), 
+#     p=2, threshold=1e8) for i, c in enumerate(base_coils + base_coils_TF)])
+# Jforce2 = LpCurveForce2(coils, coils_TF, p=2, threshold=1e8)
+# Jforce = sum([SquaredMeanForce(c, coils_TF) for c in (base_coils + base_coils_TF)])
+        
+# regularization_list = np.zeros(len(coils)) * regularization_rect(aa, bb)
+# regularization_list2 = np.zeros(len(coils_TF)) * regularization_rect(a, b)
+# Jforce = MixedLpCurveForce(coils, coils_TF, regularization_list, regularization_list2) # [SquaredMeanForce2(c, coils) for c in (base_coils)]
+# Jforce = MixedSquaredMeanForce(coils, coils_TF)
+Jforce = sum([LpCurveForce(c, coils_TF, regularization_rect(a, b), p=2, threshold=1e5 * 100) for i, c in enumerate(base_coils_TF)])
+# Jforce2 = sum([SquaredMeanForce(c, coils_TF) for c in (base_coils + base_coils_TF)])
+# Jtorque = sum([LpCurveTorque(c, coils_TF, regularization_rect(a_list[i], b_list[i]), p=2, threshold=1e5 * 40) for i, c in enumerate(base_coils + base_coils_TF)])
+# Jtorque2 = sum([SquaredMeanTorque(c, coils_TF) for c in (base_coils + base_coils_TF)])
 
-# Errors creep in when downsample = 2
-Jtorque = sum([LpCurveTorque(c, all_coils, regularization_rect(a_list[i], b_list[i]), p=2, threshold=4e5 * 100, downsample=1
-    ) for i, c in enumerate(all_base_coils)])
-Jtorque2 = sum([SquaredMeanTorque(c, all_coils, downsample=1) for c in all_base_coils])
+# Jtorque = SquaredMeanTorque2(coils, coils_TF) # [SquaredMeanForce2(c, coils) for c in (base_coils)]
+# Jtorque = [SquaredMeanTorque(c, coils_TF) for c in (base_coils + base_coils_TF)]
+
+CURVATURE_THRESHOLD = 0.5
+MSC_THRESHOLD = 0.05
+CURVATURE_WEIGHT = 1e-2
+MSC_WEIGHT = 1e-3
+Jcs = [LpCurveCurvature(c.curve, 2, CURVATURE_THRESHOLD) for c in base_coils_TF]
+Jmscs = [MeanSquaredCurvature(c.curve) for c in base_coils_TF]
 
 JF = Jf \
     + CC_WEIGHT * Jccdist \
     + CS_WEIGHT * Jcsdist \
+    + CURVATURE_WEIGHT * sum(Jcs) \
+    + MSC_WEIGHT * sum(QuadraticPenalty(J, MSC_THRESHOLD, "max") for J in Jmscs) \
     + LINK_WEIGHT * linkNum \
     + LENGTH_WEIGHT * Jlength 
 
 if FORCE_WEIGHT.value > 0.0:
     JF += FORCE_WEIGHT.value * Jforce  #\
 
-if FORCE_WEIGHT2.value > 0.0:
-    JF += FORCE_WEIGHT2.value * Jforce2  #\
+# if FORCE_WEIGHT2.value > 0.0:
+#     JF += FORCE_WEIGHT2.value * Jforce2  #\
 
-if TORQUE_WEIGHT.value > 0.0:
-    JF += TORQUE_WEIGHT * Jtorque
+# if TORQUE_WEIGHT.value > 0.0:
+#     JF += TORQUE_WEIGHT * Jtorque
 
-if TORQUE_WEIGHT2.value > 0.0:
-    JF += TORQUE_WEIGHT2 * Jtorque2
+# if TORQUE_WEIGHT2.value > 0.0:
+#     JF += TORQUE_WEIGHT2 * Jtorque2
+    # + TVE_WEIGHT * Jtve
+    # + SF_WEIGHT * Jsf
+    # + CURRENTS_WEIGHT * DipoleCurrentsObj
+    # + CURVATURE_WEIGHT * sum(Jcs_TF) \
+    # + MSC_WEIGHT * sum(QuadraticPenalty(J, MSC_THRESHOLD) for J in Jmscs_TF) \
+#    + MSC_WEIGHT * sum(QuadraticPenalty(J, MSC_THRESHOLD) for J in Jmscs) \
+    # + CURVATURE_WEIGHT * sum(Jcs) \
+
+# We don't have a general interface in SIMSOPT for optimisation problems that
+# are not in least-squares form, so we write a little wrapper function that we
+# pass directly to scipy.optimize.minimize
+
+# import pstats, io
+# from pstats import SortKey
+# print(btot.ancestors,len(btot.ancestors))
+# print(JF.ancestors,len(JF.ancestors))
+
 
 def fun(dofs):
     JF.x = dofs
+    # pr = cProfile.Profile() 
+    # pr.enable()
     J = JF.J()
     grad = JF.dJ()
+    # pr.disable()
+    # sio = io.StringIO()
+    # sortby = SortKey.CUMULATIVE
+    # ps = pstats.Stats(pr, stream=sio).sort_stats(sortby)
+    # ps.print_stats(20)
+    # print(sio.getvalue())
+    # exit()
     jf = Jf.J()
     length_val = LENGTH_WEIGHT.value * Jlength.J()
     cc_val = CC_WEIGHT * Jccdist.J()
     cs_val = CS_WEIGHT * Jcsdist.J()
     link_val = LINK_WEIGHT * linkNum.J()
     forces_val = FORCE_WEIGHT.value * Jforce.J()
-    forces_val2 = FORCE_WEIGHT2.value * Jforce2.J()
-    torques_val = TORQUE_WEIGHT.value * Jtorque.J()
-    torques_val2 = TORQUE_WEIGHT2.value * Jtorque2.J()
+    # forces_val2 = FORCE_WEIGHT2.value * Jforce2.J()
+    # torques_val = TORQUE_WEIGHT.value * Jtorque.J()
+    # torques_val2 = TORQUE_WEIGHT2.value * Jtorque2.J()
     BdotN = np.mean(np.abs(np.sum(btot.B().reshape((nphi, ntheta, 3)) * s.unitnormal(), axis=2)))
     BdotN_over_B = np.mean(np.abs(np.sum(btot.B().reshape((nphi, ntheta, 3)) * s.unitnormal(), axis=2))
         ) / np.mean(btot.AbsB())
     outstr = f"J={J:.1e}, Jf={jf:.1e}, ⟨B·n⟩={BdotN:.1e}, ⟨B·n⟩/⟨B⟩={BdotN_over_B:.1e}"
     valuestr = f"J={J:.2e}, Jf={jf:.2e}"
     cl_string = ", ".join([f"{J.J():.1f}" for J in Jls_TF])
+    kap_string = ", ".join(f"{np.max(c.kappa()):.2f}" for c in base_curves_TF)
+    msc_string = ", ".join(f"{J.J():.2f}" for J in Jmscs)
+    outstr += f", ϰ=[{kap_string}], ∫ϰ²/L=[{msc_string}]"
     outstr += f", Len=sum([{cl_string}])={sum(J.J() for J in Jls_TF):.2f}" 
     valuestr += f", LenObj={length_val:.2e}" 
     valuestr += f", ccObj={cc_val:.2e}" 
     valuestr += f", csObj={cs_val:.2e}" 
     valuestr += f", Lk1Obj={link_val:.2e}" 
     valuestr += f", forceObj={forces_val:.2e}" 
-    valuestr += f", forceObj2={forces_val2:.2e}" 
-    valuestr += f", torqueObj={torques_val:.2e}" 
-    valuestr += f", torqueObj2={torques_val2:.2e}" 
+    # valuestr += f", forceObj2={forces_val2:.2e}" 
+    # valuestr += f", torqueObj={torques_val:.2e}" 
+    # valuestr += f", torqueObj2={torques_val2:.2e}" 
     outstr += f", F={Jforce.J():.2e}"
-    outstr += f", Fnet={Jforce2.J():.2e}"
-    outstr += f", T={Jtorque.J():.2e}"
-    outstr += f", Tnet={Jtorque2.J():.2e}"
-    outstr += f", C-C-Sep={Jccdist.shortest_distance():.2f}, C-S-Sep={Jcsdist.shortest_distance():.2f}"
+    # outstr += f", Fnet={Jforce2.J():.2e}"
+    # outstr += f", T={Jtorque.J():.2e}"
+    # outstr += f", Tnet={Jtorque2.J():.2e}"
+    outstr += f", C-C-Sep={Jccdist.shortest_distance():.2f}, C-S-Sep={Jcsdist2.shortest_distance():.2f}"
     outstr += f", Link Number = {linkNum.J()}"
     outstr += f", ║∇J║={np.linalg.norm(grad):.1e}"
     print(outstr)
@@ -311,13 +369,69 @@ print("""
 ################################################################################
 """)
 
+
+print('Timing calls: ')
+t1 = time.time()
+Jf.J()
+t2 = time.time()
+print('Jf time = ', t2 - t1, ' s')
+t1 = time.time()
+Jf.dJ()
+t2 = time.time()
+print('dJf time = ', t2 - t1, ' s')
+t1 = time.time()
+Jccdist.J()
+Jccdist.dJ()
+t2 = time.time()
+print('Jcc time = ', t2 - t1, ' s')
+t1 = time.time()
+Jcsdist.J()
+Jcsdist.dJ()
+t2 = time.time()
+print('Jcs time = ', t2 - t1, ' s')
+t1 = time.time()
+linkNum.J()
+linkNum.dJ()
+t2 = time.time()
+print('linkNum time = ', t2 - t1, ' s')
+t1 = time.time()
+Jlength.J()
+Jlength.dJ()
+t2 = time.time()
+print('sum(Jls_TF) time = ', t2 - t1, ' s')
+# t1 = time.time()
+# Jforce.J()
+# t2 = time.time()
+# print('Jforces time = ', t2 - t1, ' s')
+# t1 = time.time()
+# Jforce.dJ()
+# t2 = time.time()
+# print('dJforces time = ', t2 - t1, ' s')
+# t1 = time.time()
+# Jforce2.J()
+# t2 = time.time()
+# print('Jforces2 time = ', t2 - t1, ' s')
+# t1 = time.time()
+# Jforce2.dJ()
+# t2 = time.time()
+# print('dJforces2 time = ', t2 - t1, ' s')
+# t1 = time.time()
+# Jtorque.J()
+# t2 = time.time()
+# print('Jtorques time = ', t2 - t1, ' s')
+# t1 = time.time()
+# Jtorque.dJ()
+# t2 = time.time()
+# print('dJtorques time = ', t2 - t1, ' s')
+
 n_saves = 1
-MAXITER = 500
+MAXITER = 400
 for i in range(1, n_saves + 1):
     print('Iteration ' + str(i) + ' / ' + str(n_saves))
     res = minimize(fun, dofs, jac=True, method='L-BFGS-B', 
-        options={'maxiter': MAXITER, 'maxcor': 300}, tol=1e-15)
+        options={'maxiter': MAXITER, 'maxcor': 400}, tol=1e-15)
     # dofs = res.x
+
     curves_to_vtk(
         [c.curve for c in bs_TF.coils], 
         OUT_DIR + "curves_TF_{0:d}".format(i), 
@@ -349,6 +463,6 @@ for i in range(1, n_saves + 1):
 
 t2 = time.time()
 print('Total time = ', t2 - t1)
-btot.save("biot_savart_optimized_QA.json")
+btot.save(OUT_DIR + "biot_savart_optimized_QH.json")
 print(OUT_DIR)
