+++ examples/3_Advanced/planar_coil_optimization/run_post_processing_henneberg.py
@@ -29,8 +29,8 @@ BdotN_over_B = np.mean(np.abs(np.sum(Bfield.B().reshape((nphi, ntheta, 3)) * s.u
 print(BdotN, BdotN_over_B)
 
 # # Make the QFM surfaces
-qfm_surf = make_qfm(s, Bfield)
-qfm_surf = qfm_surf.surface
+# qfm_surf = make_qfm(s, Bfield)
+# qfm_surf = qfm_surf.surface
 # qfm_surf = s
 # qfm_surf.save(filename='QA_qfm.json')
 # Bfield.set_points(qfm_surf.gamma().reshape((-1, 3)))
@@ -39,22 +39,22 @@ qfm_surf = qfm_surf.surface
 # qfm_surf.to_vtk('qfm_surf', extra_data=pointData)
 # qfm_surf.plot()
 
-from simsopt.mhd import Vmec, QuasisymmetryRatioResidual
-from simsopt.util import MpiPartition, proc0_print
+# from simsopt.mhd import Vmec, QuasisymmetryRatioResidual
+# from simsopt.util import MpiPartition, proc0_print
 
-# Run VMEC with new QFM surface
-vmec_input = "../../../tests/test_files/input.LandremanPaul2021_QA_reactorScale_lowres"
-equil = Vmec(vmec_input, mpi)
-equil.boundary = qfm_surf
-equil.run()
+# # Run VMEC with new QFM surface
+# vmec_input = "../../../tests/test_files/input.LandremanPaul2021_QA_reactorScale_lowres"
+# equil = Vmec(vmec_input, mpi)
+# equil.boundary = qfm_surf
+# equil.run()
 
-# Configure quasisymmetry objective:
-qs = QuasisymmetryRatioResidual(equil,
-                                np.arange(0, 1.01, 0.1),  # Radii to target
-                                helicity_m=1, helicity_n=0)  # (M, N) you want in |B|
+# # Configure quasisymmetry objective:
+# qs = QuasisymmetryRatioResidual(equil,
+#                                 np.arange(0, 1.01, 0.1),  # Radii to target
+#                                 helicity_m=1, helicity_n=0)  # (M, N) you want in |B|
 
-proc0_print("Quasisymmetry objective before optimization:", qs.total())
-# exit()
+# proc0_print("Quasisymmetry objective before optimization:", qs.total())
+# # exit()
 
 from simsopt.field.magneticfieldclasses import InterpolatedField
